+++ packages/orq-rc/src/orq_ai_sdk/models/findonepromptsnippetop.py
@@ -1277,7 +1277,7 @@ class FindOnePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-11T14:53:39.526Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-11T15:57:59.882Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")