+++ packages/orq-rc/src/orq_ai_sdk/prompts.py
@@ -419,12 +419,12 @@ class Prompts(BaseSDK):
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
         http_headers: Optional[Mapping[str, str]] = None,
-    ):
+    ) -> Optional[models.GetOnePromptResponseBody]:
         r"""Retrieve a prompt
 
         Retrieves a prompt object
 
-        :param id: Prompt ID
+        :param id: Unique identifier of the prompt
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
         :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
@@ -455,7 +455,7 @@ class Prompts(BaseSDK):
             request_has_path_params=True,
             request_has_query_params=True,
             user_agent_header="user-agent",
-            accept_header_value="*/*",
+            accept_header_value="application/json",
             http_headers=http_headers,
             security=self.sdk_configuration.security,
             timeout_ms=timeout_ms,
@@ -482,8 +482,10 @@ class Prompts(BaseSDK):
             retry_config=retry_config,
         )
 
-        if utils.match_response(http_res, "200", "*"):
-            return
+        if utils.match_response(http_res, "200", "application/json"):
+            return utils.unmarshal_json(
+                http_res.text, Optional[models.GetOnePromptResponseBody]
+            )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError(
@@ -512,12 +514,12 @@ class Prompts(BaseSDK):
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
         http_headers: Optional[Mapping[str, str]] = None,
-    ):
+    ) -> Optional[models.GetOnePromptResponseBody]:
         r"""Retrieve a prompt
 
         Retrieves a prompt object
 
-        :param id: Prompt ID
+        :param id: Unique identifier of the prompt
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
         :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
@@ -548,7 +550,7 @@ class Prompts(BaseSDK):
             request_has_path_params=True,
             request_has_query_params=True,
             user_agent_header="user-agent",
-            accept_header_value="*/*",
+            accept_header_value="application/json",
             http_headers=http_headers,
             security=self.sdk_configuration.security,
             timeout_ms=timeout_ms,
@@ -575,8 +577,10 @@ class Prompts(BaseSDK):
             retry_config=retry_config,
         )
 
-        if utils.match_response(http_res, "200", "*"):
-            return
+        if utils.match_response(http_res, "200", "application/json"):
+            return utils.unmarshal_json(
+                http_res.text, Optional[models.GetOnePromptResponseBody]
+            )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError(
@@ -625,7 +629,7 @@ class Prompts(BaseSDK):
     ) -> Optional[models.UpdatePromptResponseBody]:
         r"""Update a prompt
 
-        :param id: Prompt ID
+        :param id: Unique identifier of the prompt
         :param owner:
         :param domain_id:
         :param created:
@@ -774,7 +778,7 @@ class Prompts(BaseSDK):
     ) -> Optional[models.UpdatePromptResponseBody]:
         r"""Update a prompt
 
-        :param id: Prompt ID
+        :param id: Unique identifier of the prompt
         :param owner:
         :param domain_id:
         :param created:
@@ -906,7 +910,7 @@ class Prompts(BaseSDK):
     ):
         r"""Delete a prompt
 
-        :param id: Prompt ID
+        :param id: Unique identifier of the prompt
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
         :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
@@ -997,7 +1001,7 @@ class Prompts(BaseSDK):
     ):
         r"""Delete a prompt
 
-        :param id: Prompt ID
+        :param id: Unique identifier of the prompt
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
         :param timeout_ms: Override the default request timeout configuration for this method in milliseconds