+++ packages/orq-rc/RELEASES.md
@@ -409,3 +409,13 @@ Based on:
 - [python v3.1.0-rc.63] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.63] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.63 - packages/orq-rc
+
+## 2025-02-11 15:59:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.64] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.64] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.64 - packages/orq-rc